bf fixed cross-platform ping display issue across Windows and Mac

+++ src/engine/globals.h
@@ -57,8 +57,8 @@ extern int LOG_LEVEL;
 namespace network {
 // TODO add note for max name length in ui
 constexpr int MAX_NAME_LENGTH = 25;
-static long total_ping = 0;
-static long there_ping = 0;
-static long return_ping = 0;
+static long long total_ping = 0;
+static long long there_ping = 0;
+static long long return_ping = 0;
 static bool ENABLE_REMOTE_IP = false;
 }  // namespace network
