+++ src/engine/time.h
@@ -6,8 +6,8 @@
 
 namespace now {
 
-inline long current_ms() {
-    return (long) std::chrono::time_point_cast<std::chrono::milliseconds>(
+inline long long current_ms() {
+    return std::chrono::time_point_cast<std::chrono::milliseconds>(
                std::chrono::high_resolution_clock::now())
         .time_since_epoch()
         .count();
