+++ src/network/client.h
@@ -179,7 +179,7 @@ struct Client {
 
         auto update_remote_player_rare = [&](int client_id,    //
                                              int model_index,  //
-                                             int last_ping) {
+                                             long long last_ping) {
             if (!remote_players.contains(client_id)) {
                 log_warn("(rare) Remote player doesnt exist but should: {}",
                          client_id);
